+++ src/converters/piral-ng/src/startup.ts
@@ -124,13 +124,13 @@ if (process.env.NODE_ENV === 'development') {
       console.log('Running in legacy mode (Angular 2-8)');
     },
     outdated() {
-      console.log('Running in outdated mode (Angular 9-13)');
+      console.log('Running in outdated mode (Angular 9-15)');
     },
     current() {
-      console.log('Running in current mode (Angular 14-19)');
+      console.log('Running in current mode (Angular 16-20)');
     },
     next() {
-      console.log('Running in next mode (Angular 20)');
+      console.log('Running in next mode (Angular 21)');
     },
     unknown() {
       console.log('Running with an unknown version of Angular');
@@ -148,13 +148,14 @@ if (process.env.NODE_ENV === 'development') {
     v11: versionHandlers.outdated,
     v12: versionHandlers.outdated,
     v13: versionHandlers.outdated,
-    v14: versionHandlers.current,
-    v15: versionHandlers.current,
+    v14: versionHandlers.outdated,
+    v15: versionHandlers.outdated,
     v16: versionHandlers.current,
     v17: versionHandlers.current,
     v18: versionHandlers.current,
     v19: versionHandlers.current,
-    v20: versionHandlers.next,
+    v20: versionHandlers.current,
+    v21: versionHandlers.next,
   };
 
   const handler = getVersionHandler(versions) || versionHandlers.unknown;